Added token usage to the output response of the API

+++ src/api/types.rs
@@ -97,9 +97,9 @@ pub struct AssistantMessage {
 
 #[derive(Debug, Serialize, ToSchema)]
 pub struct Usage {
-    pub prompt_tokens: u32,
-    pub completion_tokens: u32,
-    pub total_tokens: u32,
+    pub prompt_tokens: u64,
+    pub completion_tokens: u64,
+    pub total_tokens: u64,
 }
 
 #[derive(Debug, Deserialize, ToSchema)]
